### 2020-03-03 #373 @akoontz11
Submission: SymbiotaR2
Submitting Author: Austin Koontz (@akoontz11 )
Repository: https://github.com/pearselab/SymbiotaR2
Version submitted: 0.0.1
Version accepted: TBD

---

```
Package: SymbiotaR2
Title: Downloading data from Symbiota2 portals into R
Version: 0.0-1
Authors@R: c(person("Austin", "Koontz", email = "austin.koontz@usu.edu", role = c("aut", "cre")),
    person("William D.", "Pearse", email = "will.pearse@gmail.com", role ="aut", comment=c(ORCID="0000-0002-6241-3164")))
Maintainer: Austin Koontz <austin.koontz@usu.edu>
Description: Download data from Symbiota2 portals using Symbiota's
    API. Covers the Checklists, Collections, Crowdsource, Exsiccati,
    Glossary, ImageProcessor, Key, Media, Occurrence, Reference, Taxa,
    Traits, and UserRoles API families. Each Symbiota2 portal owner
    can load their own plugins (and modified code), and so this
    package may not cover every possible API endpoint from a given
    Symbiota2 instance.
LazyLoad: yes
    knitr (>= 1.23),
    vcr (>= 0.3.0)
    xml2 (>= 1.2.0),
    rjson (>= 0.2.20),
    fs (>= 1.2.6)
```

	- [ ] workflow automataion

The primary purpose of SymbiotaR2 is to allow users to pull objects from the specified Symbiota2 database portal into an R environment. It does this by wrapping each of the default API calls made available on the Symbiota2 portal of interest.

The target audiences includes anyone interested in using R to analyze biological specimen data made available through the Symbiota2 content management system. Potential scientific applications include comparing biological trait data between different lineages, analyzing phenological data, and biodiversity research generally.

No R packages have been developed for the [updated code structure provided by Symbiota2](http://symbiota.org/docs/project-summary/).
