### 2020-03-03 #373 @akoontz11
Tagging @willpearse as a contributing author of this package as well.
