### 2020-03-13 #373 @noamross
### Editor checks:

---

#### Editor comments

Thank you for your submission @akoontz11!

Here is output from `goodpractice::gp()`. Your test coverage is good, and the >80 lines are mostly just due to indention conventins.  I note that you use `sapply()` in basically the same way all over the package and it might make sense to make a small utility function replacing this use that uses `vapply()`, which might also make code clearer.  I also note that the documentation can be a little terse; README or other documentation might go into more detail, as to the the nature of Symbiota data and data types following the [principle of mutiple points of entry](https://devguide.ropensci.org/building.html#documentation). None of these things require holding up review; I am seeking reviewers and will ask them to evaluate these observations.

```
── GP SymbiotaR2 ────────────────────────────

It is good practice to

    R/Defaults.R:28:NA
    R/Defaults.R:31:NA
    R/Defaults.R:32:NA
    R/Defaults.R:33:NA
    R/Defaults.R:34:NA
    ... and 57 more lines

    R/Checklists.R:3:1
    R/Checklists.R:10:1
    R/Checklists.R:11:1
    R/Checklists.R:21:1
    R/Checklists.R:43:1
    ... and 99 more lines

    vapply() instead.

    R/Checklists.R:24:13
    R/Checklists.R:32:20
    R/Checklists.R:70:17
    R/Checklists.R:90:13
    R/Checklists.R:98:20
    ... and 14 more lines
```
---

Reviewers:
Due date:
