### 2020-03-14 #373 @willpearse
Brilliant, thank you Dr Ross! If it's OK with you, can we wait until after
the reviews are back in to address your `sapply` and documentation
suggestions? I imagine there will be other comments and it might be easier
to address everything at once.

>
