### 2020-06-04 #373 @willpearse
Thanks Noam. I do know some, but I worry that suggesting one of my friends
would be a bit unethical! Have you found this list (
http://symbiota.org/docs/symbiota-introduction/active-symbiota-projects/)?
I should add that many herbarium managers are not R developers, and so they
might be unwilling to review an R package (they're often taxonomic
experts). Another option might be someone who has spent a lot of time
working with natural history collection data; an example would be Emily
Meinke (although we shared a post-doc advisor; https://emilykmeineke.com/).

>
