### 2020-06-15 #373 @bpbond
@willpearse @akoontz11 What's the workflow for running the package tests? I've tried `devtools::test()` and source'ing `tests/test-all.R`, and get errors both ways. Thanks.
