### 2020-06-16 #373 @akoontz11
Hi @bpbond, thank you for agreeing to review this package!

Here's the workflow we've used for installation and testing:
```
Rscript -e "require(roxygen2);roxygenize('~/Symbiota2/code/SymbiotaR2/')"
R CMD build ~/Symbiota2/code/SymbiotaR2
Rscript -e "install.packages('SymbiotaR2_0.0.1-tar.gz')"
R CMD check --as-cran SymbiotaR2_0.0-1.tar.gz
```
Could you let me know what errors you're getting, using either `devtools::test("SymbiotaR2")`, `test_package("SymbiotaR2")`, or source'ing `test-all.R`? I had to call `test_check("SymbiotaR2")` from within the `tests` directory, in order to get that method to work. We're using Travis-CI and `covr` to report test coverage, which is ~87%.
