### 2020-07-08 #373 @bpbond
Using `R CMD check` as above does run the tests successfully. I can't get `test_check()` to work regardless of the working directory. Anyway, thanks, and more in a bit.
