### 2020-07-09 #373 @bpbond
Apologies for the delay.

- [X] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work

>##### For packages co-submitting to JOSS
>
>- [X] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [X] **A short summary** describing the high-level functionality of the software
>- [X] **Authors:**  A list of authors with their affiliations
>- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Biodiversity researchers and institutions face formidable challenges in integrating biological community knowledge and data, and this package aims to provide an easy-to-use interface to [Symbiota2](https://symbiota2.org). Thus I appreciate (after some research) the goals of this package, and the work the authors have put into it.

That said, it seems like there's a fair amount of work that could be done on this package to make it more accessible for beginners and non-specialists (or specialists unfamiliar with R packages), increase its robustness, and generally conform better to best practices. Many of the limitations below made it impossible for me, as a reviewer, to fully test or understand the package's functionality.

The comments below, which I hope are useful for the authors, are grouped into broad categories for readability:

#### Building

R CMD check` gives the following note:

```
The Title field should be in title case. Current version is:
‘Downloading data from Symbiota2 portals into R’
```

Package dependencies are minimal and sensible.

I highly recommend that @akoontz11 get an [ORCID ID](https://orcid.org) and add it to the DESCRIPTION file.

The package's documentation root could be made much more useful and welcoming. Here it is:
![Screen Shot 2020-07-09 at 5 39 05 AM](https://user-images.githubusercontent.com/1956468/87027730-99ac5d80-c1ab-11ea-81d7-827ea95785d1.png)

By way of comparison, here is the documentation home for the [zoo](https://cran.r-project.org/web/packages/zoo/index.html) package:
![Screen Shot 2020-07-09 at 5 38 53 AM](https://user-images.githubusercontent.com/1956468/87027741-9d3fe480-c1ab-11ea-9748-d308c993ff9b.png)

The latter provides easy-to-find links to documentation, demos, vignettes, and NEWS. I'd recommend adopting this as a model for Symbiota2.

WHAT IS SYMBIOTA2? The README and main package help pages have _nothing_ about the overall context behind, and need for, this package. The vignette gives slightly more background info but still not enough. I suggest that the authors include something like this (adapted from the Symbiota2 website):
>Symbiota is open source software that enables US natural history collections to share information from their collections in ways that benefit multiple user groups.

The package documentation should reference both the Symbiota2 website and the Gries et al. 2014 paper linked therein.

The vignette is quite bare-bones. In addition, its purported output is statically pasted into the `Rmd` source file, i.e. it's not really being run. OK, but when I tried to run the very first code example in the vignette, got an error than the URL couldn't be reached. As a beginning user, now what do I do?

Most function do not have provided examples.

#### Code quality

The package code is generally clean, clearly and consistently formatted, and easy to follow.

There's no defensive programming that I can see, e.g. assertions on function entry or exit. This might be a useful addition.

I appreciate that testing network-oriented functions is a challenge. It would be useful if the package test code itself had bit of documentation, in particular information about the testing strategy and what's going on with the YAML files in the `tests/testthat` folder.
