### 2020-07-13 #373 @akoontz11
Hi @bpbond, thanks very much for taking the time to put together this review and for your feedback! I agree with everything mentioned here: the package documentation could be more informative, and I think that providing documentation for the testing code is an excellent idea, given the importance of those tests for users trying to connect to a portal. I think these changes will greatly improve the package.

If possible, it'd be helpful for me to receive feedback from the other reviewer before incorporating these changes, so that I could make all suggested changes at once. Thank you again, for this.
