### 2020-08-13 #373 @sckott
- there's no examples other than for the setup fxn. I'd normally say there should be egs for every exported function, but if there's no public symbiota instance to access then that makes including examples hard.
    - ideally make a .github dir and include in there contributing.md, as well as issue and pull request templates

>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).
>.  - the 1 reference that probably has a DOI does not include a DOI

    - tests do not run for me. some suggestions:
        - from an organization perspective, its a good idea to put your vcr cassettes into their own folder, and tests in their own folder. if you run `vcr::use_vcr()` it will set things up in the suggested way with the cassettes in `tests/fixtures` and test files in `test/testhat`.  https://github.com/ropensci/vcr#getting-started
        - probably best to namespace the `use_cassette` calls, so use `vcr::use_cassette` - running tests errors for me as R doesn't know where to find `use_cassette`
        - to get the most accurate information on what lines errors occur on in tests, better to wrap just the functions that make http requests in use_cassette and then the testthat expectations below that - optional of course
    - see comments below

---

- agree with 1st reviewer that the pkg would greatly benefit from parameter type checking - e.g., `assert(some_parameter, "character")`
- the functions throughout the pkg use `NA` as default. this isn't covered per se in ropensci guidelines, but this strikes me as out of the ordinary. often fxns have default as missing or `NULL`, with functions to check those things within the function, e.g., `missing(param)` or `is.null(param)`.
- there's a mix of styles with parameters and functions. for parameters, some are all lowercase, some all uppercase. making all lowercase would make sense, and follow the guidelines. functions are consistent though, all being TitleCase
- maintainers could make their work easier using `@inheritParams` or `@template` to reduce redundant parameter definitions
- is it true that a user using this pkg would likely only use one url per R session? if so, maybe it would make sense to put the `url` parameter as the last one in each function as the user is not likely to set the value of that parameter, so they could use more succinct code by only setting the id and/or page
- with `SymbiotaR2_setup()` i can set any arbitrary URL that exists. can the url check be changed to make sure the url is for a symbiota instance?
- are there any public symbiota instances? if so that'd be useful to make it possible for users to try out the package more easily
- strongly recommend building the readme.md from a readme.rmd and building the vignette from actual code - you'd probably want to prebuild the vignette for this pkg though, see e.g., [this post](https://ropensci.org/technotes/2019/12/08/precompute-vignettes/)
- code style: lots of lines in the package > 80 width, not super important, but improves readability if lines are 80 char or less
- authors mention no other R packages - could they comment on https://github.com/FranzKrah/rSymbiota and what overlap if any?
- no top level manual file exists - per guidelines: "The package should contain top-level documentation for ?foobar"
