### 2020-08-26 #373 @akoontz11
Hi @sckott and @bpbond, thanks very much for all of the helpful feedback on SymbiotaR2! I'm working on each of these points now and will post in this thread once these changes have been incorporated.
