### 2020-10-08 pearselab/SymbiotaR2#9 @akoontz11
Building package documentation from code
#### Updates pending release of public Symbiota2 instance
- Package documentation needs to be updated to build the README.md file from a .Rmd.
- Prebuild package vignette (addressing portal connectivity using knitr; see [here](https://ropensci.org/technotes/2019/12/08/precompute-vignettes/))
