### 2020-10-08 #373 @akoontz11
@bpbond and @sckott: firstly, thank you again for reviewing SymbiotaR2 and providing all of the helpful feedback! I've copied each of your review comments and pasted them below, along with responses on points where you've recommended changes (headed with "AK:"). For some of these points, I've included explanations as to why we haven't addressed the issue.

Please let me know if there's anything I can do to clarify any of my points below, and again, thanks for contributing your help.

--------------------------------------------------------------------------------------------------------------------------------------------------------------
@bpbond
--------------------------------------------------------------------------------------------------------------------------------------------------------------
> Apologies for the delay.
>
> ## Package Review
> * **Briefly describe any working relationship you have (had) with the package authors.**
> * [x]  As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work
>
> #### Documentation
> The package includes all the following forms of documentation:
>
> * [ ]  **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
> AK: The[ README has been updated](https://github.com/pearselab/SymbiotaR2/blob/master/README.md) to provide a statement of need for SymbiotaR2.
> * [x]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
> * [x]  **Vignette(s)** demonstrating major functionality that runs successfully locally
> * [x]  **Function Documentation:** for all exported functions in R help
> * [ ]  **Examples** for all exported functions in R Help that run successfully locally
> * [ ]  **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).
> AK: A CONTRIBUTING.md file has been provided within a [.github directory](https://github.com/pearselab/SymbiotaR2/tree/master/.github)
>
> > ##### For packages co-submitting to JOSS
> >
> > * [x]  The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
> >
> > The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
> >
> > * [x]  **A short summary** describing the high-level functionality of the software
> > * [x]  **Authors:**  A list of authors with their affiliations
> > * [x]  **A statement of need** clearly stating problems the software is designed to solve and its target audience.
> > * [ ]  **References:** with DOIs for all those that have one (e.g. papers, datasets, software).
>
> #### Functionality
> * [x]  **Installation:** Installation succeeds as documented.
> * [ ]  **Functionality:** Any functional claims of the software been confirmed.
> * [ ]  **Performance:** Any performance claims of the software been confirmed.
> * [x]  **Automated tests:** Unit tests cover essential functions of the package
>   and a reasonable range of inputs and conditions. All tests pass on the local machine.
> * [x]  **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines
>
> #### Final approval (post-review)
> * [ ]  **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**
>
> Estimated hours spent reviewing: 2
>
> * [ ]  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.
>
> ### Review Comments
> Biodiversity researchers and institutions face formidable challenges in integrating biological community knowledge and data, and this package aims to provide an easy-to-use interface to [Symbiota2](https://symbiota2.org). Thus I appreciate (after some research) the goals of this package, and the work the authors have put into it.
>
> That said, it seems like there's a fair amount of work that could be done on this package to make it more accessible for beginners and non-specialists (or specialists unfamiliar with R packages), increase its robustness, and generally conform better to best practices. Many of the limitations below made it impossible for me, as a reviewer, to fully test or understand the package's functionality.
>
> The comments below, which I hope are useful for the authors, are grouped into broad categories for readability:
>
> #### Building
> `R CMD check` gives the following note:
>
> ```
> The Title field should be in title case. Current version is:
> ‘Downloading data from Symbiota2 portals into R’
> ```
> AK: Thanks for this, @bpbond  --the Title field had been updated to [use title case](https://github.com/pearselab/SymbiotaR2/blob/b8ba7f49f1d3ccba738fbb14dc7a61b61ca12b2b/DESCRIPTION#L3)
>
> Package dependencies are minimal and sensible.
>
> I highly recommend that @akoontz11 get an [ORCID ID](https://orcid.org) and add it to the DESCRIPTION file.
>
> AK: @akoontz11's ORCID ID has been added to the [DESCRIPTION file](https://github.com/pearselab/SymbiotaR2/blob/b8ba7f49f1d3ccba738fbb14dc7a61b61ca12b2b/DESCRIPTION#L5).
>
> #### Documentation
> The package's documentation root could be made much more useful and welcoming. Here it is:
> ![Screen Shot 2020-07-09 at 5 39 05 AM](https://user-images.githubusercontent.com/1956468/87027730-99ac5d80-c1ab-11ea-81d7-827ea95785d1.png)
>
> By way of comparison, here is the documentation home for the [zoo](https://cran.r-project.org/web/packages/zoo/index.html) package:
> ![Screen Shot 2020-07-09 at 5 38 53 AM](https://user-images.githubusercontent.com/1956468/87027741-9d3fe480-c1ab-11ea-9748-d308c993ff9b.png)
>
> The latter provides easy-to-find links to documentation, demos, vignettes, and NEWS. I'd recommend adopting this as a model for Symbiota2.
>
> AK: Thanks for this recommendation. The documentation head has been updated to include a link to the vignette, and [a NEWS.md file ](https://github.com/pearselab/SymbiotaR2/blob/master/NEWS.md)has been included in the root directory.
>
> WHAT IS SYMBIOTA2? The README and main package help pages have _nothing_ about the overall context behind, and need for, this package. The vignette gives slightly more background info but still not enough. I suggest that the authors include something like this (adapted from the Symbiota2 website):
>
> > Symbiota is open source software that enables US natural history collections to share information from their collections in ways that benefit multiple user groups.
>
> The package documentation should reference both the Symbiota2 website and the Gries et al. 2014 paper linked therein.
>
>
> The vignette is quite bare-bones. In addition, its purported output is statically pasted into the `Rmd` source file, i.e. it's not really being run. OK, but when I tried to run the very first code example in the vignette, got an error than the URL couldn't be reached. As a beginning user, now what do I do?
>
> AK: Thanks for this, @bpbond. The README has been updated to provide more context for SymbiotaR2, and an `R/SymbiotaR2-package.R` file has also been included to provide more information within the package documentation. The vignette has also been updated to include more sections (similar to the README). Both include links to the Symbiota website and the Gries et al. 2014 paper. Finally, the vignette includes some description of the package code, as well as a [Troubleshooting section](https://github.com/pearselab/SymbiotaR2/blob/b8ba7f49f1d3ccba738fbb14dc7a61b61ca12b2b/vignettes/SymbiotaR2-intro.Rmd#L168).
> Because Symbiota2 is still in development, there currently isn't a public-facing portal we could direct users to, so we can't provide a URL that users can work with right now. We appreciate and agree that this is far from ideal, but because we're trying to get everything 'feature complete' for the launch of Symbiota2, we're working towards having the R package ready, citable, and online for when existing portals upgrade to the new version of Symbiota.
>
> Most function do not have provided examples.
>
> AK: This has been updated, and examples are now included in the help files, with one for each API family of functions.
>
> #### Code quality
> The package code is generally clean, clearly and consistently formatted, and easy to follow.
>
> There's no defensive programming that I can see, e.g. assertions on function entry or exit. This might be a useful addition.
>
> AK: Thanks for recommending this. Parameter type checking has been incorporated for several of the package's root functions (i.e. [`.check.url`](https://github.com/pearselab/SymbiotaR2/blob/b8ba7f49f1d3ccba738fbb14dc7a61b61ca12b2b/R/utils.R#L73) and [`.api.scaffold`](https://github.com/pearselab/SymbiotaR2/blob/b8ba7f49f1d3ccba738fbb14dc7a61b61ca12b2b/R/utils.R#L38)--but please note that `id` arguments do not use parameter type checking, as some Symbiota2 API handles utilize a non-numeric `id`)
>
> I appreciate that testing network-oriented functions is a challenge. It would be useful if the package test code itself had bit of documentation, in particular information about the testing strategy and what's going on with the YAML files in the `tests/testthat` folder.
>
> AK: I totally agree, I think there's a clear value in being more explicit regarding the testing. To address this, Testing sections have been included in both the README and [the vignette](https://github.com/pearselab/SymbiotaR2/blob/b8ba7f49f1d3ccba738fbb14dc7a61b61ca12b2b/vignettes/SymbiotaR2-intro.Rmd#L190). There are some additional complexities to this beyond network-oriented testing (see above the comment about the chicken-and-egg problem of portals not yet having upgraded to Symbiota2).
--------------------------------------------------------------------------------------------------------------------------------------------------------------

@sckott
--------------------------------------------------------------------------------------------------------------------------------------------------------------
> ## Package Review
> * **Briefly describe any working relationship you have (had) with the package authors.**
> * [x]  As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).
>
> #### Documentation
> The package includes all the following forms of documentation:
>
> * [ ]  **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
> AK: The README has been updated to provide a statement of need for SymbiotaR2.
> * [x]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
> * [x]  **Vignette(s)** demonstrating major functionality that runs successfully locally
> * [x]  **Function Documentation:** for all exported functions in R help
> * [ ]  **Examples** for all exported functions in R Help that run successfully locally
>
>   * there's no examples other than for the setup fxn. I'd normally say there should be egs for every exported function, but if there's no public symbiota instance to access then that makes including examples hard.
>
> AK: Examples have now been included for each API family of SymbiotaR2 functions (see example [here](https://github.com/pearselab/SymbiotaR2/blob/b8ba7f49f1d3ccba738fbb14dc7a61b61ca12b2b/R/Checklists.R#L12)). That being said, we definitely agree that a public Symbiota2 instance would be really useful for explaining function outputs--but unfortunately, there's no instance we could reference for that, right now (see above).
>
> * [ ]  **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).
>
>   * ideally make a .github dir and include in there contributing.md, as well as issue and pull request templates
>
> AK: Thanks for this suggestion, this is very helpful. A .github directory [has been added to the root folder](https://github.com/pearselab/SymbiotaR2/tree/master/.github), and includes a CONTRIBUTING.md file
>
> > ##### For packages co-submitting to JOSS
> >
> > * [x]  The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
> >
> > The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
> >
> > * [x]  **A short summary** describing the high-level functionality of the software
> > * [x]  **Authors:**  A list of authors with their affiliations
> > * [x]  **A statement of need** clearly stating problems the software is designed to solve and its target audience.
> > * [ ]  **References:** with DOIs for all those that have one (e.g. papers, datasets, software).
> >   .  - the 1 reference that probably has a DOI does not include a DOI
>
> AK: Good catch: [a DOI has been included](https://github.com/pearselab/SymbiotaR2/blob/b8ba7f49f1d3ccba738fbb14dc7a61b61ca12b2b/paper.bib#L18) for the Gries et al. 2014 paper.
>
> #### Functionality
> * [x]  **Installation:** Installation succeeds as documented.
> * [ ]  **Functionality:** Any functional claims of the software been confirmed.
> * [ ]  **Performance:** Any performance claims of the software been confirmed.
> * [ ]  **Automated tests:** Unit tests cover essential functions of the package
>   and a reasonable range of inputs and conditions. All tests pass on the local machine.
>
>   * tests do not run for me. some suggestions:
>
>     * from an organization perspective, its a good idea to put your vcr cassettes into their own folder, and tests in their own folder. if you run `vcr::use_vcr()` it will set things up in the suggested way with the cassettes in `tests/fixtures` and test files in `test/testhat`.  https://github.com/ropensci/vcr#getting-started
>
> AK: Thanks for detailing this, @sckott. The `tests` directory [has been updated to follow this format](https://github.com/pearselab/SymbiotaR2/tree/master/tests), and a [`helper-SymbiotaR2.R` file](https://github.com/pearselab/SymbiotaR2/blob/master/tests/testthat/helper-SymbiotaR2.R) has been included in the `tests/testthat` directory.
>
>     * probably best to namespace the `use_cassette` calls, so use `vcr::use_cassette` - running tests errors for me as R doesn't know where to find `use_cassette`
>
> AK: Great point. All `use_cassette` commands are now namespaced
>
>     * to get the most accurate information on what lines errors occur on in tests, better to wrap just the functions that make http requests in use_cassette and then the testthat expectations below that - optional of course
>
> AK: Thanks for this recommendation, as this should make future troubleshooting easier. Tests have been updated to only wrap SymbiotaR2 commands in `use_cassette` (example [here](https://github.com/pearselab/SymbiotaR2/blob/b8ba7f49f1d3ccba738fbb14dc7a61b61ca12b2b/tests/testthat/test_Checklist.R#L4))
>
> * [ ]  **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines
>
>   * see comments below
>
> #### Final approval (post-review)
> * [ ]  **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**
>
> Estimated hours spent reviewing: 1.5
>
> * [ ]  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.
>
> ### Review Comments
> * agree with 1st reviewer that the pkg would greatly benefit from parameter type checking - e.g., `assert(some_parameter, "character")`
>
> AK: See response above: parameter type checking incorporated for [`.check.url`](https://github.com/pearselab/SymbiotaR2/blob/b8ba7f49f1d3ccba738fbb14dc7a61b61ca12b2b/R/utils.R#L73) and [`.api.scaffold`](https://github.com/pearselab/SymbiotaR2/blob/b8ba7f49f1d3ccba738fbb14dc7a61b61ca12b2b/R/utils.R#L38) functions (but `id` parameters are not checked to be numeric, as a few functions use a non-numeric `id` argument).
>
> * the functions throughout the pkg use `NA` as default. this isn't covered per se in ropensci guidelines, but this strikes me as out of the ordinary. often fxns have default as missing or `NULL`, with functions to check those things within the function, e.g., `missing(param)` or `is.null(param)`.
>
> AK: This has been updated: missing arguments now utilize `NULL` as the default, rather than `NA` (example [here](https://github.com/pearselab/SymbiotaR2/blob/b8ba7f49f1d3ccba738fbb14dc7a61b61ca12b2b/R/utils.R#L13))
>
> * there's a mix of styles with parameters and functions. for parameters, some are all lowercase, some all uppercase. making all lowercase would make sense, and follow the guidelines. functions are consistent though, all being TitleCase
>
> AK: Thanks for calling this out. All function parameters are now in lowercase. Function names remain unchanged.
>
> * maintainers could make their work easier using `@inheritParams` or `@template` to reduce redundant parameter definitions
>
> AK: Thank you so much for this suggestion! `@template` [is now being used](https://github.com/pearselab/SymbiotaR2/blob/master/man-roxygen/SymbiotaR2.R) to build help pages for most package functions.
>
> * is it true that a user using this pkg would likely only use one url per R session? if so, maybe it would make sense to put the `url` parameter as the last one in each function as the user is not likely to set the value of that parameter, so they could use more succinct code by only setting the id and/or page
>
> AK: Great point--`url` has been moved to be the last argument specified for all functions.
>
> * with `SymbiotaR2_setup()` i can set any arbitrary URL that exists. can the url check be changed to make sure the url is for a symbiota instance?
>
> AK: Good point. To address this, we have incorporated an API call at the end of the [`.check.url` function](https://github.com/pearselab/SymbiotaR2/blob/b8ba7f49f1d3ccba738fbb14dc7a61b61ca12b2b/R/utils.R#L83) (call for `Occurrences(id=1)`).
>
> * are there any public symbiota instances? if so that'd be useful to make it possible for users to try out the package more easily
>
> AK: Thanks for this recommentation, @sckott. We agree that public instance of Symbiota2 would greatly help to simplify the SymbiotaR2 package documentation (and would be generally helpful for Symbiota2 usage). While this is something we've mentioned to the development team, there currently isn't a public Symbiota2 instance available. As we outline above, we are trying to build the infrastructure that supports Symbitoa2 portals at the same time as the software itself, so that when we have an official "launch day" managers can upgrade their portals and everything will "just work". We appreciate this makes things much more difficult for the reviewers and we do apologize about that!
>
> * strongly recommend building the readme.md from a readme.rmd and building the vignette from actual code - you'd probably want to prebuild the vignette for this pkg though, see e.g., [this post](https://ropensci.org/technotes/2019/12/08/precompute-vignettes/)
>
[…]
