### 2020-11-17 #373 @noamross
My apologies for missing this last exchange and letting this linger.  Thank you @sckott and @bpbond for your thourough reviews and @akoontz11 for your response.  @bpbond and @sckott, can you please look at @akoontz11's replies and us know if they address the concerns in your review.
