### 2020-11-17 #373 @bpbond
I haven't actually installed and tested the revised package (should I?), but @akoontz11 's responses above address all my concerns. Very comprehensive and convincing.
