### 2020-12-01 #373 @akoontz11
Hi @noamross, thanks for this checklist, the guidance, and the approval! I believe I've made most of the necessary updates you've listed, but let me know if anything looks astray.

@bpbond and @sckott, I would like to include you as `rev` authors for the package--if that is alright with you, let me know and I will updated the DESCRIPTION file. Once I do, I can tag and create a release for Zenodo.
